src/requirements.txt
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.4
decorator==5.1.1
Deprecated==1.2.14
fabric==3.2.2
idna==3.4
invoke==2.2.0
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
paramiko==3.3.1
pexpect==4.9.0
psycopg2==2.9.9
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.16.1
PyNaCl==1.5.0
pypsexec==0.3.0
pyspnego==0.10.2
python-dotenv==1.0.0
regex==2023.8.8
requests==2.31.0
rich==13.5.2
smbprotocol==1.13.0
tiktoken==0.4.0
urllib3==2.0.6
wrapt==1.15.0
